Initial restructure of representatives section (now university section)
