+++ src/views/ContentView.tsx
@@ -1,10 +1,9 @@
 import HomeView from './HomeView';
 import GroupsView from './GroupsView';
-import NewsView from './NewsView';
 import OrganizationView from './OrganizationView';
 import RulesView from './RulesView';
 import ServicesView from './ServicesView';
-import RepresentativesView from './RepresentativesView';
+import UniversityView from './UniversityView';
 import { Route, Switch } from 'react-router-dom';
 import { withCookies } from 'react-cookie';
 
@@ -15,11 +14,10 @@ const ContentView = () => {
     return (
         <Switch>
             <Route path="/home" render={() => wrap(<HomeView/>)} />
-            <Route path="/rules" render={() => wrap(<RulesView/>)} />
-            <Route path='/news' render={() => wrap(<NewsView/>)} />
             <Route path='/courses' render={() => wrap(<GroupsView/>)} />
             <Route path='/services' render={() => wrap(<ServicesView/>)} />
-            <Route path='/representatives' render={() => wrap(<RepresentativesView/>)} />
+            <Route path="/rules" render={() => wrap(<RulesView/>)} />
+            <Route path='/representatives' render={() => wrap(<UniversityView/>)} />
             <Route path='/organization' render={() => wrap(<OrganizationView/>)} />
         </Switch>
     )
